changed PDD type to decimal
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C29319

+++ QualityMeasurement/Code/QM.CLITester.Entities/CLITesterConnectorBase.cs
@@ -86,7 +86,7 @@ namespace QM.CLITester.Entities
         public int CliSuccess { get; set; }
         public int CliNoResult { get; set; }
         public int CliFail { get; set; }
-        public int Pdd { get; set; }
+        public Decimal Pdd { get; set; }
         public string ShareUrl { get; set; }
     }
 }
